refactor(frontend): clarify editor property helper naming
