+++ apps/frontend/src/components/editor/EditorPropertyControls.tsx
